requirements.txt
appnope==0.1.2
backcall==0.2.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
decorator==5.1.0
distlib==0.3.4
filelock==3.4.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-MySQLdb==0.2.0
Flask-PyMongo==2.3.0
ipython==7.30.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.27
mysqlclient==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
protobuf==3.19.1
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.11.1
PyJWT==2.3.0
pymongo==4.0.1
PyMySQL==1.0.2
python-dotenv==0.19.2
six==1.16.0
traitlets==5.1.1
virtualenv==20.12.1
wcwidth==0.2.5
Werkzeug==2.0.2
